[FIX] Duplicated Server Initialization
