+++ src/init-server.test.ts
@@ -4,6 +4,7 @@ import { StdioServerTransport } from '@modelcontextprotocol/sdk/server/stdio.js'
 import { afterEach, beforeEach, describe, expect, it, type Mock, vi } from 'vitest'
 import { resolveCredentialState } from './credential-state.js'
 import { initServer } from './init-server.js'
+import { createMcpServer } from './server-factory.js'
 import { loadConfig } from './tools/helpers/config.js'
 import { ensureValidToken } from './tools/helpers/oauth2.js'
 import { registerTools } from './tools/registry.js'
@@ -26,6 +27,10 @@ vi.mock('./tools/helpers/oauth2.js', () => ({
 }))
 vi.mock('./tools/registry.js', () => ({
   registerTools: vi.fn()
+}))
+vi.mock('./server-factory.js', () => ({
+  createMcpServer: vi.fn(),
+  getVersion: vi.fn()
 }))
 vi.mock('@modelcontextprotocol/sdk/server/index.js', () => ({
   Server: vi.fn()
@@ -61,12 +66,15 @@ describe('initServer', () => {
     vi.mocked(readFileSync).mockReturnValue(JSON.stringify({ name: '@n24q02m/better-email-mcp', version: '1.0.0' }))
 
     // Default mock implementation for Server
+    const mockServer = {
+      connect: vi.fn().mockResolvedValue(undefined)
+    }
     // biome-ignore lint/complexity/useArrowFunction: Must use function for constructor mocking
     ;(Server as unknown as Mock).mockImplementation(function () {
-      return {
-        connect: vi.fn().mockResolvedValue(undefined)
-      }
+      return mockServer
     })
+
+    vi.mocked(createMcpServer).mockReturnValue(mockServer as any)
 
     // Default mock implementation for StdioServerTransport
     // biome-ignore lint/complexity/useArrowFunction: Must use function for constructor mocking
@@ -90,46 +98,11 @@ describe('initServer', () => {
 
     // Verify
     expect(loadConfig).toHaveBeenCalled()
-    expect(Server).toHaveBeenCalledWith(
-      expect.objectContaining({
-        name: '@n24q02m/better-email-mcp'
-      }),
-      expect.any(Object)
-    )
-    expect(registerTools).toHaveBeenCalledWith(server, mockAccounts)
+    expect(createMcpServer).toHaveBeenCalledWith(mockAccounts)
     expect(StdioServerTransport).toHaveBeenCalled()
     expect(server!.connect).toHaveBeenCalledWith(expect.anything())
   })
 
-  it('uses fallback version 0.0.0 when package.json not found', async () => {
-    const mockAccounts = [{ email: 'test@example.com' }]
-    vi.mocked(loadConfig).mockReturnValue(mockAccounts as any)
-    vi.mocked(existsSync).mockReturnValue(false)
-
-    await initServer()
-
-    expect(Server).toHaveBeenCalledWith(
-      expect.objectContaining({
-        version: '0.0.0'
-      }),
-      expect.any(Object)
-    )
-  })
-
-  it('uses fallback version when package.json has no version field', async () => {
-    const mockAccounts = [{ email: 'test@example.com' }]
-    vi.mocked(loadConfig).mockReturnValue(mockAccounts as any)
-    vi.mocked(readFileSync).mockReturnValue(JSON.stringify({ name: '@n24q02m/better-email-mcp' }))
-
-    await initServer()
-
-    expect(Server).toHaveBeenCalledWith(
-      expect.objectContaining({
-        version: '0.0.0'
-      }),
-      expect.any(Object)
-    )
-  })
 
   it('starts server with warning when credentials set but no accounts loaded', async () => {
     // Setup mocks -- resolveCredentialState returns configured but loadConfig returns empty
@@ -142,7 +115,7 @@ describe('initServer', () => {
     // Verify server still starts
     expect(loadConfig).toHaveBeenCalled()
     expect(consoleSpy).toHaveBeenCalledWith('Warning: No email accounts configured')
-    expect(Server).toHaveBeenCalled()
+    expect(createMcpServer).toHaveBeenCalled()
     expect(exitSpy).not.toHaveBeenCalled()
   })
 
@@ -223,13 +196,11 @@ describe('initServer', () => {
     await initServer()
 
     expect(resolveCredentialState).toHaveBeenCalledTimes(1)
-    expect(Server).toHaveBeenCalled()
+    expect(createMcpServer).toHaveBeenCalledWith([])
     expect(consoleSpy).toHaveBeenCalledWith(
       'Server starting without credentials. Tools will guide setup on first call.'
     )
     // loadConfig should NOT be called when not configured
     expect(loadConfig).not.toHaveBeenCalled()
-    // registerTools called with empty accounts
-    expect(registerTools).toHaveBeenCalledWith(expect.anything(), [])
   })
 })
